asistente_ia_atletas_project/requirements.txt
streamlit
mediapipe
opencv-python
pillow
openai
